requirements.txt
cython>=0.29.27
flask>=2.0.1
flask_caching>=1.10.1
flask-cors>=3.0.10
flask_restful>=0.3.9
gevent>=21.8.0
gql>=3.0.0 
requests-toolbelt>=0.9.0
webargs>=8.0.1
Werkzeug>=2.0.1
web3>=5.11.1